requirements.txt
# 核心依赖
requests>=2.32.3
python-dotenv>=1.0.1
APScheduler>=3.10.4
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
